+++ src/main/java/org/sparky/sparkyai/usercall/dto/UserCallResponseDto.java
@@ -9,7 +9,7 @@ public class UserCallResponseDto {
 
     private String prompt;
     private String response;
-    private Long consumedTokens;
+    private Integer consumedTokens;
     private ZonedDateTime createdAt;
 
 }
